embedding/requirements.txt
pyyaml
tenacity
openai
qdrant-client
supabase
python-dotenv